+++ examples/tron_send_transaction.py
@@ -1,5 +1,6 @@
-from aiotx.clients import AioTxTRONClient
 import asyncio
+
+from aiotx.clients import AioTxTRONClient
 
 tron_client = AioTxTRONClient(node_url="https://api.shasta.trongrid.io")