+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.10.1]
+- timeout_between_blocks optional param added to block monitoring
+- optional headers param added to all client 
+- add jsonRPC into TON url by default and Deprication warning added
+
 ## [2.9.1]
 - remove not tested python versions from pypi list